chore: remove unnecessary or broken tests, move authCredentials to models folder
