+++ src/app/features/sign-up/sign-up.store-facade.ts
@@ -2,7 +2,7 @@ import { inject, Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { selectErrorMessage } from '../../core/states/error/error.selectors';
 import { resolveError } from '../../core/states/error/error.actions';
-import { SignUpCredentials } from '../../shared/types/authCredentials.type';
+import { SignUpCredentials } from '../../core/models/authCredentials.interface';
 import { signUp } from '../../core/states/auth/auth.actions';
 
 @Injectable({ providedIn: 'root' })
